worker/requirements.txt
Flask==2.3.0
redis==4.5.0
Pillow==9.5.0
requests==2.31.0
prometheus_client==0.20.0
psycopg[binary]==3.1.18
